Add OGC proxy for WMS/WFS
This adds a rack proxy at /ogc that simply proxies WMS and WFS requests
to GeoServer on the backend. For users who have authenticated the
configured BasicAuth username and password are added to the request to
GeoServer.

The short answer for why this is necessary is that it prevents us from
having to run two GeoServer instances. The longer answer is forthcoming
in a future ADR.