Adds support for developer auth and saml auth
Developer auth is super helpful for local development, PR builds and possibly staging dependent on how much auth flows matter for the app (i.e. do you need to test this in a prod-like environment before prod... if so, use the same auth as prod).

SAML auth is likely a good option for production.

Shibboleth / mod_shib support is retained for legacy reasons.