Enables multiple concurrent authentication types
This commit is not intended to be merged into master, but instead be used as a temporary solution to help in our transition from exisiting servers with shib to new servers with saml. A lot of additional things are changing with the new servers so we want to test the new auth on the old servers before moving forward.

ENV['AUTH_TYPE'] must be set to one ore more (comma separated) of:
- shibboleth
- saml
- developer

All can be active, developer should NOT be included for any public facing systems with real data loaded.

Only one link will be provided in the header for login with the preference order of shibboleth, saml, developer. All other routes exist and can be accessed manually by maniuplating the URL.