Fix Touchstone authentication
Why these changes are being introduced:

The _user_util_links partial is not where Blacklight expects it to be,
breaking Touchstone auth.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/UXWS-1121

How this addresses that need:

This moves the offending partial to where it needs to be (up one directory).

Side effects of this change:

None, hopefully.